Advanced enemy generation_2
Now each spawn point creates its own type of enemies and has a target, that is, an object to which the enemies go. At the moment of creation, the spawn point gives its target to the enemy so that he heads towards it. There should be several such goals, each spawner has its own goal.

The targets themselves move along a predetermined route, from one point to another.